Add replay parser written in Zig
there are lots of different parsers in many languages, including python,
C#, Java, JavaScript, Rust (this one even has a python bindings),
but all of them are not suitable for our needs.

With our own parser we can parse and skip whatever we want, whenever we
want, and do not depend on third party implementations.
(which also means that we do not have to do any post-processing of
the data in slow python)